requirements.txt
anyio==4.6.2.post1
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
Django==5.1.2
djangorestframework==3.15.2
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
psycopg2==2.9.10
python-telegram-bot==21.6
requests==2.32.3
sniffio==1.3.1
sqlparse==0.5.1
urllib3==2.2.3
